Basic Stopping Criteria and Criticality Loops
This commit implements stopping via `AbstractStoppingCriterion`.
A stopping criterion is checkd pre or post iteration, post
criticality computation and pre or post a criticality loop.

The Criticality Routine is implemented according to the theory.
We introduce a new cache object and train models on a
decreased radius.
To keep old models, new methods `copy_model` and `copyto_model!`
are introduced.
The database handling in RBFModels was changed to reflect the
needs of a criticality routine.

The signature of some functions (especially in "descent.jl")
has changed. I am not sure about this.

There is also a meta data object, the contents of which depend a
bit about the current position in the algorithm.
We now have so many caches and containers that I am tempted to
forgo most of the container structs in favor of separate objects
and very verbose method signatures...